Implement A13 messages. Re-enable test case.

+++ hl7-reader/src/main/java/uk/ac/ucl/rits/inform/pipeline/informdb/PatientFactRepository.java
@@ -9,5 +9,5 @@ import uk.ac.ucl.rits.inform.informdb.PatientFact;
  *
  * @author Jeremy Stein
  */
-public interface PatientFactRepository extends CrudRepository<PatientFact, Integer> {
+public interface PatientFactRepository extends CrudRepository<PatientFact, Long> {
 }
